refactor: exercices.composition > Composition.exercise
