+++ _05-OOP-2_polymorphism/src/Composition/exercise/Bed.java
@@ -1,4 +1,4 @@
-package exercices.composition;
+package Composition.exercise;
 
 public class Bed {
     private String style;
