+++ _05-OOP-2_polymorphism/src/Composition/exercise/Lamp.java
@@ -1,4 +1,4 @@
-package exercices.composition;
+package Composition.exercise;
 
 public class Lamp {
     private String style;
